Move embed-builder into a separate subfolder
