+++ commands/active.js
@@ -5,17 +5,13 @@ const logger = pino({
   },
 });
 const { SlashCommandBuilder } = require("discord.js");
-const { EmbedBuilder } = require("discord.js");
-const totalWorldList = require("../data/total-worlds.json");
-const activeStarsEmbed = require("../utils/active-stars-embed.js");
-const db = require("../utils/db.js");
-const { format, formatDistanceToNow, parseISO } = require("date-fns");
+const activeStarsEmbed = require("../embeds/active-stars-embed.js");
 
 const data = new SlashCommandBuilder()
   .setName("active")
   .setDescription("Get a list of released stars");
 
-async function run({ c, client, interaction }) {
+async function run({ interaction }) {
   logger.info("/active");
 
   // Defer the reply to ensure enough time to process the command
